@@ -1,3 +1,3 @@
 module github.com/michaelsatish/junit-xml-viewer
 
-go 1.18
+go 1.21
